Added write binary gltf stage
