+++ lib/parseBinaryGltf.js
@@ -9,6 +9,7 @@ module.exports = parseBinaryGltf;
 function parseBinaryGltf(data) {
     var sizeOfUint32 = Uint32Array.BYTES_PER_ELEMENT;
     var byteOffset = 0;
+    // console.log(data.slice(0, 5*sizeOfUint32));
 
     if (data.slice(byteOffset, sizeOfUint32).toString() !== 'glTF') {
         throw new Error('File is not valid binary glTF');
@@ -22,6 +23,7 @@ function parseBinaryGltf(data) {
     var decoder = new StringDecoder();
     var scene = decoder.write(data.slice(byteOffset, byteOffset + sceneLength));
     var gltf = JSON.parse(scene);
+    // console.log(data.slice(byteOffset, byteOffset + sceneLength));
 
     byteOffset += sceneLength;
     var body = data.slice(byteOffset);
