+++ lib/writeGltf.js
@@ -1,11 +1,9 @@
 'use strict';
 var fs = require('fs');
 var path = require('path');
-var Cesium = require('cesium');
-var defined = Cesium.defined;
-var dataUri = require('datauri');
 var async = require('async');
 var mkdirp = require('mkdirp');
+var writeSource = require('./writeSource');
 
 module.exports = writeGltf;
 
@@ -15,10 +13,10 @@ function writeGltf(gltf, outputPath, isEmbedded, createDirectory, callback) {
         outputPath = path.join(path.dirname(outputPath), 'output', path.basename(outputPath));
         mkdirp.sync(path.dirname(outputPath));
     }
-    var filePath = path.dirname(outputPath);
+    var basePath = path.dirname(outputPath);
 
     async.each(['buffers', 'images', 'shaders'], function(name, asyncCallback) {
-            writeSource(gltf, filePath, name, isEmbedded, asyncCallback);
+            writeSource(gltf, basePath, name, isEmbedded, asyncCallback);
         }, function(err) {
             if (err) {
                 if (callback) {
@@ -40,62 +38,4 @@ function writeGltf(gltf, outputPath, isEmbedded, createDirectory, callback) {
             }   
         }
     );
-}
-
-function writeSource(gltf, filePath, name, isEmbedded, callback) {
-    var objects = gltf[name];
-
-    //Iterate through each object and write out its uri
-    if (defined(objects)) {
-        var ids = Object.keys(objects);
-        async.each(ids, function(id, asyncCallback) {
-            var object = objects[id];
-            if (defined(object.extras) && defined(object.extras,source)) {
-                var source = object.extras.source;
-                var extension = object.extras.extension;
-                delete object.extras.source;
-                delete object.extras.extension;
-
-                //Write the source object as a data or file uri depending on the isEmbedded flag
-                if (isEmbedded) {
-                    var sourceDataUri = new dataUri();
-                    if (name === 'shaders') {
-                        sourceDataUri.format('.txt', source);
-                    }
-                    else {
-                        sourceDataUri.format(extension, source);
-                    }
-                    object.uri = sourceDataUri.content;
-                    process.nextTick(function() {
-                        asyncCallback();
-                    });
-                }
-                else {
-                    var fileName = id + extension;
-                    object.uri = fileName;
-                    var outputPath = path.join(filePath, fileName);
-                    fs.writeFile(outputPath, source, function (err) {
-                        asyncCallback(err);
-                    });  
-                }
-            }
-        }, function(err) {
-            if (err) {
-                if (callback) {
-                    callback(err);
-                }
-                else{
-                    throw err;
-                }
-            }
-            else if (callback) {
-                callback();
-            }
-        });
-    }
-    else {
-        if (callback) {
-            callback();
-        }
-    }
 }
